refactor(components): colocate tests alongside implementations
Move 8 test files from __tests__/ directories into their respective
component/util directories and update import paths. Fix pre-existing
lint errors in JsonForm test (unnecessary act wrapping).
